backend/pyproject.toml
[tool.poetry]
name = "cryptopatro"
version = "0.1.0"
description = "Personal portfolio tracker and trading assistant bot"
authors = ["Doruk Gezici <doruk@gezici.me>"]
readme = "README.md"

[tool.poetry.dependencies]
binance-connector = "^3.7.0"
celery = "^5.4.0"
django-celery-beat = "^2.6.0"
django-celery-results = "^2.5.1"
django-cors-headers = "^4.3.1"
django-ninja = "^1.1.0"
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
python-telegram-bot = "^21.3"
sentry-sdk = "^2.5.1"
uvicorn = "^0.30.1"
whitenoise = "^6.6.0"
python = "^3.12"

[tool.poetry.group.dev.dependencies]
celery-types = "^0.22.0"
django-stubs = "^5.0.2"
ruff = "^0.4.9"
taskipy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
dev = "uvicorn cryptopatro.asgi:application --host 0.0.0.0 --port 8000 --reload"
lint = "ruff check"
format = "ruff format"
export = "poetry export --output requirements.txt"

[tool.ruff]
line-length = 120
indent-width = 4

[tool.ruff.lint]
select = ["E", "F"]
ignore = ["F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
docstring-code-line-length = 120
